Force deploy for each project
After the work in #128 to stop crawling staging deploys, we need to force a deploy for each product in order to have the robots.txt go live. This commit adds a blank file to each product to force it to deploy. It will be removed from master once all the deploys have successfully run once with the new robots.txt logic.