[1.1.1.1] Fix cloudflared proxy-dns system service instructions (#442)
`cloudflared service install` does not work on Linux (TUN-3520) when
trying to configure a proxy-dns service. Aside from that, the extra
auto-updater service installed by that command might not be desirable
either. Suggest the installation of an unprivileged systemd service with
the correct dependencies to ensure that it starts as early as possible
to avoid DNS resolution failures, or fallback to other nameservers in
/etc/resolv.conf.

Also suggest modifying `/etc/resolv.conf`, that should almost always
work. Specific Linux distributions can provide instructions for their
platforms.