Update plugin-compatibility.md (#9212)
I am the co-developer of the Perfmatters plugin. Our plugin fully supports Cloudflare, and we recommend it to all of our clients. In fact, around 95% of them use Cloudflare, including ourselves. 

Many clients have asked why we aren't on the compatible list. So getting on here would save everyone time.