Restructure get-started to notify of split local/remote KV's before we document accessing from Workers (#15235)
When you follow this guide, you 1. create a worker project, 2. create a remote kv namespace via wrangler/dashboard, 3.a interact with wrangler, 3.b then create a worker to interact to with the kv account. 

However, 3.b guides the reader to write code, and test it out. It leads the reader to believe they will be able to retrieve the values they have added to the kv namespace in 3.a. This is misleading, and it is only covered in 4. develop locally from wrangler. In that section, it is mentioned that `wrangler dev` specifically uses a different, local-only version of KV from `wrangler kv key get...`

By separating accessing KV from a Worker and elevating the warning of remote vs local KV before the code snippets, readers can avoid being misled into trying to access remote resources from local wrangler.