Update delete-account.mdx (#24102)
On this Cloudflare document, it says - After you delete your profile, you cannot use the email address used with your profile to create a new account. 

However, after checking Cloudflare wiki, Our macro for account deletion, Cloudflare community posts, and confirming in the Cloudflare IAM chat, I realized that this is not true and it is confusing a lot of customers.

I changed this sentence with the one that explicitly mentions that for most user deletes, the email will be freed up to be used in a new signup right away, except for users who have a lock on their account (for legal reasons).

I want to propose this change so that the users can clarify there doubts regarding account deletion by themselves and don't have to make multiple support cases and wait in queues for it.

Thank you.