DS-15647 DS-15648: Document aggregate-if and date/time function additions to the WAE SQL API (#26101)
* DS-15647 DS-15648: Document aggregate-if and date/time function additions to the WAE SQL API

This commit adds the following function additions to the docs:

- `countIf`
- `sumIf`
- `avgIf`
- `toYear`
- `toMonth`
- `toDayOfMonth`
- `toHour`
- `toMinute`
- `toSecond`
- `toStartOfYear`
- `toStartOfMonth`
- `toStartOfDay`
- `toStartOfHour`
- `toStartOfFifteenMinutes`
- `toStartOfMinute`
- `today`

This commit reorders the date/time docs page to be in approximate alphabetical
order, which works much better than just appending the functions to the end.

* DS-15648: Document further date/time function additions to the WAE SQL API

These are some assorted functions commonly used by LLMs:

  - `toDayOfWeek`
  - `toStartOfWeek`
  - `toStartOfTenMinutes`
  - `toStartOfFiveMinutes`
  - `toYYYYMM`

* Add changelog for further WAE SQL additions

This commit adds a changelog entry for recent function additions,
as documented in the previous commits.

I have chosen to omit descriptions of the date/time functions,
because they are all quite generic and it looks unwieldy to have
some many samey descriptions.

* Update src/content/changelog/workers/2025-11-12-analytics-engine-further-sql-enhancements.mdx

Co-authored-by: Brendan Irvine-Broque <brendanib@gmail.com>

---------

Co-authored-by: Brendan Irvine-Broque <brendanib@gmail.com>